+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
-## [Unreleased]
+## [22.0.11-1.12] - 2025-01-10
 ### Fixed
 - Handling of changed profile information during sign-up when review profile is set to "missing" or T&C is requested
 - Add value to lastrefresh in saml federation only after SAML federation is finished adding all IdPs